requirements.txt
black==22.6.0
pylint==2.14.5
pytest==7.1.2
pytest-runner==6.0.0
websocket-client==1.3.3
websockets==10.3
requests==2.28.1
rich==12.5.1
pandas==1.4.3
numpy==1.23.2
matplotlib==3.3.3